Move scipy-imported files to scipy/ subdir
