+++ .pre-commit-config.yaml
@@ -4,4 +4,4 @@ repos:
     hooks:
     - id: black
       language_version: python3.6
-      exclude: 'utils/scipy_'
+      exclude: 'utils/scipy/'
