+++ corrscope/utils/scipy/signal.py
