+++ corrscope/utils/scipy/wavfile.py
