+++ corrscope/utils/scipy/windows.py
